backend/requirements.txt
# Core framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.0.0

# HTTP client
httpx==0.26.0
requests==2.31.0

# Testing tools
pytest==7.4.3
pytest-mock==3.14.0
pytest-asyncio==0.21.0

# Libraries
python-dotenv==1.0.0
rapidfuzz==3.0.0

# MongoDB
# motor==3.3.2
# pymongo==4.6.2